feat(packages): update package.json files for public access
This commit updates the package.json files of several packages. The main change is the addition of a "publishConfig" property, which sets the access level to public. This allows these packages to be easily and publicly accessible when published. The lint scripts have also been slightly adjusted for improved consistency across packages.